:recycle: refactor code (#26)
