+++ excore/__init__.py
@@ -1,9 +1,18 @@
-from . import hub
+from . import config, hub
 from ._constants import __author__, __version__, _load_workspace_config, _workspace_cfg
-from .config import ConfigArgumentHookProtocol, build_all, load, set_target_fields
-from .hook import Hook, HookManager
-from .logger import _enable_excore_debug, add_logger, init_logger, remove_logger
-from .registry import Registry, load_registries
+from .config.config import build_all, load
+from .config.parse import set_target_fields
+from .engine import hook, logging, registry
+from .engine.hook import ConfigArgumentHook, Hook, HookManager
+from .engine.logging import (
+    _enable_excore_debug,
+    add_logger,
+    debug_only,
+    init_logger,
+    logger,
+    remove_logger,
+)
+from .engine.registry import Registry, load_registries
 
 __all__ = [
     "__author__",
@@ -13,7 +22,7 @@ __all__ = [
     "hook",
     "logger",
     "registry",
-    "ConfigArgumentHookProtocol",
+    "ConfigArgumentHook",
     "build_all",
     "load",
     "Hook",
@@ -22,6 +31,8 @@ __all__ = [
     "remove_logger",
     "Registry",
     "load_registries",
+    "debug_only",
+    "logging",
 ]
 
 init_logger()
